ai-api/requirements.txt
fastapi
uvicorn[standard]
gunicorn
pydantic
pydantic-settings
sqlmodel
cytomine-python-client
opencv-python-headless
tifffile
randomcolor
